feat: treat css as entry
BREAKING CHANGE: until now, a CSS file was always coupled to a JS
entry: `vite()->css('index.js')` and the JS entry would have to import
the CSS file. With this change, the CSS file is an entry itself and can
be loaded by its name: `vite()->css('index.css')`.
Note: for this to work the CSS file has to be added to
`rollupOptions.input` in the `vite.config.js`.